@@ -1,5 +1,7 @@
 package com.coreoz.plume.jersey.errors;
 
+import lombok.extern.slf4j.Slf4j;
+
 import jakarta.ws.rs.WebApplicationException;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
@@ -7,33 +9,27 @@ import jakarta.ws.rs.core.Response.Status;
 import jakarta.ws.rs.ext.ExceptionMapper;
 import jakarta.ws.rs.ext.Provider;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import java.util.List;
 
-import com.google.common.collect.ImmutableList;
-
+@Slf4j
 @Provider
 public class WsResultExceptionMapper implements ExceptionMapper<Throwable> {
-
-	private static final Logger logger = LoggerFactory.getLogger(WsResultExceptionMapper.class);
-
 	@Override
 	public Response toResponse(Throwable e) {
-		if (e instanceof WsException) {
-			WsException wsException = (WsException) e;
-			return Response
+		if (e instanceof WsException wsException) {
+            return Response
 				.status(Status.BAD_REQUEST)
 				.entity(new ErrorResponse(wsException.getError(), wsException.getStatusArguments()))
 				.type(MediaType.APPLICATION_JSON_TYPE)
 				.build();
 		}
-		if(e instanceof WebApplicationException) {
-			return ((WebApplicationException) e).getResponse();
+		if(e instanceof WebApplicationException webApplicationException) {
+			return webApplicationException.getResponse();
 		}
 		if(e instanceof JsonRequestParseException) {
 			return Response
 				.status(Status.BAD_REQUEST)
-				.entity(new ErrorResponse(WsError.REQUEST_INVALID, ImmutableList.of("JSON object supplied in request is invalid")))
+				.entity(new ErrorResponse(WsError.REQUEST_INVALID, List.of("JSON object supplied in request is invalid")))
 				.type(MediaType.APPLICATION_JSON_TYPE)
 				.build();
 		}
@@ -42,9 +38,8 @@ public class WsResultExceptionMapper implements ExceptionMapper<Throwable> {
 
 		return Response
 			.status(Status.INTERNAL_SERVER_ERROR)
-			.entity(new ErrorResponse(WsError.INTERNAL_ERROR, ImmutableList.of()))
+			.entity(new ErrorResponse(WsError.INTERNAL_ERROR, List.of()))
 			.type(MediaType.APPLICATION_JSON_TYPE)
 			.build();
 	}
-
 }
