@@ -1,17 +1,16 @@
 package com.coreoz.plume.mail;
 
+import com.typesafe.config.Config;
+import com.typesafe.config.ConfigFactory;
 import org.assertj.core.api.Assertions;
 import org.junit.Test;
 
-import com.google.common.collect.ImmutableMap;
-import com.typesafe.config.Config;
-import com.typesafe.config.ConfigFactory;
+import java.util.Map;
 
 public class MailerProviderTest {
-
 	@Test
 	public void should_generate_well_form_property_file() {
-		Config config = ConfigFactory.parseMap(ImmutableMap.of(
+		Config config = ConfigFactory.parseMap(Map.of(
 			"mail.\"javaxmail.debug\"", "true",
 			"mail.transportstrategy", "SMTP_SSL"
 		));
@@ -22,5 +21,4 @@ public class MailerProviderTest {
 			+ "simplejavamail.transportstrategy=SMTP_SSL"
 		);
 	}
-
 }
