@@ -1,17 +1,14 @@
 package com.coreoz.plume.jersey.grizzly;
 
+import lombok.extern.slf4j.Slf4j;
 import org.glassfish.grizzly.http.server.ErrorPageGenerator;
 import org.glassfish.grizzly.http.server.Request;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * Replace the default Grizzly error handler to avoid leaking server information
  */
+@Slf4j
 public class GrizzlyErrorPageHandler implements ErrorPageGenerator {
-
-	private static final Logger logger = LoggerFactory.getLogger(GrizzlyErrorPageHandler.class);
-
 	@Override
 	public String generate(Request request, int status, String reasonPhrase, String description,
 			Throwable exception) {
@@ -27,5 +24,4 @@ public class GrizzlyErrorPageHandler implements ErrorPageGenerator {
 
 		return status + " " + reasonPhrase;
 	}
-
 }
