@@ -1,18 +1,14 @@
 package com.coreoz.plume.jersey.java8;
 
-import java.time.LocalDateTime;
+import com.google.common.base.Strings;
+import lombok.extern.slf4j.Slf4j;
 
 import jakarta.ws.rs.ext.ParamConverter;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import java.time.LocalDateTime;
 
-import com.google.common.base.Strings;
-
+@Slf4j
 public class LocalDateTimeConverter implements ParamConverter<LocalDateTime> {
-
-	private static final Logger logger = LoggerFactory.getLogger(LocalDateTimeConverter.class);
-
 	@Override
 	public LocalDateTime fromString(String value) {
 		try {
@@ -27,5 +23,4 @@ public class LocalDateTimeConverter implements ParamConverter<LocalDateTime> {
 	public String toString(LocalDateTime value) {
 		return value == null ? null : value.toString();
 	}
-
 }
