@@ -1,24 +1,23 @@
 package com.coreoz.plume.jersey.jee;
 
-import java.io.IOException;
-import java.io.PrintWriter;
+import lombok.extern.slf4j.Slf4j;
 
-import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
+import java.io.PrintWriter;
+
+@Slf4j
 public class ErrorServlet extends HttpServlet {
-
-	private static final long serialVersionUID = 6186826368888103404L;
-
 	@Override
-	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+	protected void doGet(HttpServletRequest req, HttpServletResponse resp) {
 		try(PrintWriter printWriter = resp.getWriter()) {
 			int errorCode = resp.getStatus();
 			String errorMessage = HttpStatus.getStatusText(errorCode);
 			printWriter.println("Error " + errorCode + " : " + errorMessage);
-		}
+		} catch (Exception e) {
+            logger.error("Error displaying error message for error code {}", resp.getStatus());
+        }
 	}
-
 }
