@@ -1,8 +1,6 @@
 package com.coreoz.plume.jersey.security.basic;
 
-import java.nio.charset.StandardCharsets;
-import java.util.Base64;
-import java.util.function.Function;
+import lombok.extern.slf4j.Slf4j;
 
 import jakarta.ws.rs.ClientErrorException;
 import jakarta.ws.rs.ForbiddenException;
@@ -11,18 +9,17 @@ import jakarta.ws.rs.core.HttpHeaders;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.core.Response.Status;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import java.nio.charset.StandardCharsets;
+import java.util.Base64;
+import java.util.function.Function;
 
 /**
  * Enable to easily manage HTTP Basic Authentication with Jersey.
  *
  * @param <U> The user type handled by the authentication
  */
+@Slf4j
 public class BasicAuthenticator<U> {
-
-	private static final Logger logger = LoggerFactory.getLogger(BasicAuthenticator.class);
-
 	private static final String BASIC_PREFIX = "Basic ";
 
 	private final Function<Credentials, U> authenticator;
@@ -136,5 +133,4 @@ public class BasicAuthenticator<U> {
 
 		return new Credentials(decodedCredentials[0], decodedCredentials[1]);
 	}
-
 }
